Fix typo (Verlog => Verilog)
Once I saw it, I could not unsee it. It chewed at my soul and I had to fix it.

On another note, it may be helpful to have each sentence on its own line. Assuming that the website formats the text similarly to markdown, this would be beneficial because it appears the same when rendered but breaks up long lines, making it easier to notice changes within modified lines later on.